Fix dynamic minimum fan table issue caused by python3 (#6690)
**- Why I did it**
After migrating to python3, the operator '/' always get a float result, but it gets integer result in python2. Need fix this in thermal_conditions.

**- How I did it**
1. cast float value to int
2. change the unit test case to cover this situation

**- How to verify it**
Manually test and regression test

+++ platform/mellanox/mlnx-platform-api/sonic_platform/thermal_conditions.py
@@ -85,7 +85,7 @@ class MinCoolingLevelChangeCondition(ThermalPolicyConditionBase):
 
         trust_state = Thermal.check_module_temperature_trustable()
         temperature = Thermal.get_min_amb_temperature()
-        temperature = temperature / 1000
+        temperature = int(temperature / 1000)
 
         change_cooling_level = False
         if trust_state != MinCoolingLevelChangeCondition.trust_state:
